+++ app/src/main/java/sample/kingja/loadsir/MainActivity.java
@@ -5,6 +5,11 @@ import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 
+import sample.kingja.loadsir.target.AnimateActivity;
+import sample.kingja.loadsir.target.BestPracticesActivity;
+import sample.kingja.loadsir.target.DefaultCallbackActivity;
+import sample.kingja.loadsir.target.KeepTitleActivity;
+import sample.kingja.loadsir.target.KeepTitleFragmentActivity;
 import sample.kingja.loadsir.target.MultiFragmentActivity;
 import sample.kingja.loadsir.target.FragmentSingleActivity;
 import sample.kingja.loadsir.target.NormalActivity;
@@ -32,6 +37,10 @@ public class MainActivity extends AppCompatActivity {
         startActivity(new Intent(this, NormalActivity.class));
     }
 
+    public void showPlaceholder(View view) {
+        startActivity(new Intent(this, PlaceholderActivity.class));
+    }
+
     public void inActivityWithConvertor(View view) {
         startActivity(new Intent(this, ConvertorActivity.class));
     }
@@ -40,20 +49,35 @@ public class MainActivity extends AppCompatActivity {
         startActivity(new Intent(this, FragmentSingleActivity.class));
     }
 
-    public void inView(View view) {
-        startActivity(new Intent(this,ViewTargetActivity.class));
+    public void inFragmentMutil(View view) {
+        startActivity(new Intent(this, MultiFragmentActivity.class));
     }
 
     public void inFragmentViewSirPager(View view) {
-        startActivity(new Intent(this,MultiFragmentWithViewPagerActivity.class));
+        startActivity(new Intent(this, MultiFragmentWithViewPagerActivity.class));
     }
 
-
-    public void inFragmentMutil(View view) {
-        startActivity(new Intent(this,MultiFragmentActivity.class));
+    public void inView(View view) {
+        startActivity(new Intent(this, ViewTargetActivity.class));
     }
 
-    public void showPlaceholder(View view) {
-        startActivity(new Intent(this,PlaceholderActivity.class));
+    public void defaultCallback(View view) {
+        startActivity(new Intent(this, DefaultCallbackActivity.class));
+    }
+
+    public void animatCallback(View view) {
+        startActivity(new Intent(this, AnimateActivity.class));
+    }
+
+    public void keepTitleInFragment(View view) {
+        startActivity(new Intent(this, KeepTitleFragmentActivity.class));
+    }
+
+    public void bestPractices(View view) {
+        startActivity(new Intent(this, BestPracticesActivity.class));
+    }
+
+    public void keepTitleInActivity(View view) {
+        startActivity(new Intent(this, KeepTitleActivity.class));
     }
 }