+++ app/src/main/java/sample/kingja/loadsir/callback/CustomCallback.java
@@ -2,7 +2,6 @@ package sample.kingja.loadsir.callback;
 
 import android.content.Context;
 import android.view.View;
-import android.widget.ImageView;
 import android.widget.Toast;
 
 import com.kingja.loadsir.callback.Callback;
@@ -23,8 +22,8 @@ public class CustomCallback extends Callback {
     }
 
     @Override
-    protected boolean onRetry(final Context context, View view) {
-        Toast.makeText(context.getApplicationContext(), "Hello mother fuck! :p", Toast.LENGTH_SHORT).show();
+    protected boolean onReloadEvent(final Context context, View view) {
+        Toast.makeText(context.getApplicationContext(), "Hello buddy, how r u! :p", Toast.LENGTH_SHORT).show();
         (view.findViewById(R.id.iv_gift)).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {