+++ app/src/main/java/sample/kingja/loadsir/callback/LoadingCallback.java
@@ -1,5 +1,8 @@
 package sample.kingja.loadsir.callback;
 
+import android.content.Context;
+import android.view.View;
+
 import com.kingja.loadsir.callback.Callback;
 
 import sample.kingja.loadsir.R;
@@ -16,5 +19,15 @@ public class LoadingCallback extends Callback {
     @Override
     protected int onCreateView() {
         return R.layout.layout_loading;
+    }
+
+    @Override
+    public boolean getSuccessVisible() {
+        return super.getSuccessVisible();
+    }
+
+    @Override
+    protected boolean onReloadEvent(Context context, View view) {
+        return true;
     }
 }