+++ app/src/main/java/sample/kingja/loadsir/target/NormalActivity.java
@@ -1,18 +1,20 @@
 package sample.kingja.loadsir.target;
 
+import android.content.Context;
 import android.os.Bundle;
 import android.os.SystemClock;
 import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
-
-import com.kingja.loadsir.callback.Callback;
+import android.widget.TextView;
 
 import sample.kingja.loadsir.callback.EmptyCallback;
 import sample.kingja.loadsir.callback.LoadingCallback;
 
+import com.kingja.loadsir.callback.Callback;
 import com.kingja.loadsir.core.LoadService;
 import com.kingja.loadsir.core.LoadSir;
+import com.kingja.loadsir.core.Transport;
 
 import sample.kingja.loadsir.R;
 import sample.kingja.loadsir.PostUtil;
@@ -26,7 +28,6 @@ import sample.kingja.loadsir.PostUtil;
  */
 
 public class NormalActivity extends AppCompatActivity {
-
 
     private LoadService loadService;
 
@@ -50,8 +51,14 @@ public class NormalActivity extends AppCompatActivity {
                     }
                 }).start();
             }
+        }).setCallBack(EmptyCallback.class, new Transport() {
+            @Override
+            public void order(Context context, View view) {
+                TextView mTvEmpty = (TextView) view.findViewById(R.id.tv_empty);
+                mTvEmpty.setText("fine, no data. You must fill it!");
+            }
         });
-        PostUtil.postCallbackDelayed(loadService,EmptyCallback.class);
+        PostUtil.postCallbackDelayed(loadService, EmptyCallback.class);
     }
 
 }