+++ loadsir/src/main/java/com/kingja/loadsir/core/LoadLayout.java
@@ -2,10 +2,12 @@ package com.kingja.loadsir.core;
 
 import android.content.Context;
 import android.support.annotation.NonNull;
+import android.view.View;
 import android.widget.FrameLayout;
 
 import com.kingja.loadsir.LoadSirUtil;
 import com.kingja.loadsir.callback.Callback;
+import com.kingja.loadsir.callback.SuccessCallback;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -17,10 +19,12 @@ import java.util.Map;
  * Email:kingjavip@gmail.com
  */
 
-class LoadLayout extends FrameLayout {
+public class LoadLayout extends FrameLayout {
     private Map<Class<? extends Callback>, Callback> callbacks = new HashMap<>();
     private Context context;
     private Callback.OnReloadListener onReloadListener;
+    private Class<? extends Callback> preCallback;
+    private static final int CALLBACK_CUSTOM_INDEX = 1;
 
     public LoadLayout(@NonNull Context context) {
         super(context);
@@ -30,6 +34,13 @@ class LoadLayout extends FrameLayout {
         this(context);
         this.context = context;
         this.onReloadListener = onReloadListener;
+    }
+
+    public void setupSuccessLayout(Callback callback) {
+        addCallback(callback);
+        View successView = callback.getRootView();
+        successView.setVisibility(View.GONE);
+        addView(successView);
     }
 
     public void setupCallback(Callback callback) {
@@ -45,10 +56,7 @@ class LoadLayout extends FrameLayout {
     }
 
     public void showCallback(final Class<? extends Callback> callback) {
-        if (!callbacks.containsKey(callback)) {
-            throw new IllegalArgumentException(String.format("The Callback (%s) is nonexistent.", callback
-                    .getSimpleName()));
-        }
+        checkCallbackExist(callback);
         if (LoadSirUtil.isMainThread()) {
             showCallbackView(callback);
         } else {
@@ -66,13 +74,44 @@ class LoadLayout extends FrameLayout {
     }
 
     private void showCallbackView(Class<? extends Callback> status) {
-        if (getChildCount() > 0) {
-            removeAllViews();
+        if (preCallback != null) {
+            if (preCallback == status) {
+                return;
+            }
+            callbacks.get(preCallback).onDetach();
         }
+        if (getChildCount() > 1) {
+            removeViewAt(CALLBACK_CUSTOM_INDEX);
+        }
+
         for (Class key : callbacks.keySet()) {
             if (key == status) {
-                addView(callbacks.get(key).getRootView());
+                SuccessCallback successCallback = (SuccessCallback) callbacks.get(SuccessCallback.class);
+                if (key == SuccessCallback.class) {
+                    successCallback.show();
+                } else {
+                    successCallback.showWithCallback(callbacks.get(key).getSuccessVisible());
+                    View rootView = callbacks.get(key).getRootView();
+                    addView(rootView);
+                    callbacks.get(key).onAttach(context, rootView);
+                }
+                preCallback = status;
             }
+        }
+    }
+
+    public void setCallBack(Class<? extends Callback> callback, Transport transport) {
+        if (transport == null) {
+            return;
+        }
+        checkCallbackExist(callback);
+        transport.order(context, callbacks.get(callback).obtainRootView());
+    }
+
+    private void checkCallbackExist(Class<? extends Callback> callback) {
+        if (!callbacks.containsKey(callback)) {
+            throw new IllegalArgumentException(String.format("The Callback (%s) is nonexistent.", callback
+                    .getSimpleName()));
         }
     }
 }